requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.5
django-heroku==0.3.1
djangorestframework==3.15.2
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.15
langchain-core==0.3.31
langchain-text-splitters==0.3.5
langsmith==0.2.11
multidict==6.1.0
numpy==2.2.2
orjson==3.10.15
packaging==24.2
propcache==0.2.1
psycopg2==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.8.2
yarl==1.18.3
